Updated the notification publisher so the logs are displayed as a separate UI + tweaks to sending messages
